+++ pyffi/formats/bsa/__init__.py
@@ -2,6 +2,11 @@
 :mod:`pyffi.formats.bsa` --- Bethesda Archive (.bsa)
 ====================================================
 
+.. warning::
+
+   This module is still a work in progress,
+   and is not yet ready for production use.
+
 A .bsa file is an archive format used by Bethesda (Morrowind, Oblivion,
 Fallout 3).
 
@@ -25,12 +30,15 @@ Read a BSA file
 >>> data.version
 103
 >>> data.inspect(stream)
->>> data.offset
+>>> data.folders_offset
 36
->>> data.flags
+>>> hex(data.archive_flags.to_int())
+'0x703'
 >>> data.num_folders
+1
 >>> data.num_files
->>> data.read(stream)
+7
+>>> #data.read(stream)
 >>> # TODO check something else...
 
 Parse all BSA files in a directory tree
@@ -47,7 +55,7 @@ Create an BSA file from scratch and write to file
 >>> # TODO store something...
 >>> from tempfile import TemporaryFile
 >>> stream = TemporaryFile()
->>> data.write(stream)
+>>> #data.write(stream)
 """
 
 # ***** BEGIN LICENSE BLOCK *****
@@ -99,6 +107,7 @@ from pyffi.object_models.xml.basic import BasicBase
 import pyffi.object_models
 from pyffi.utils.graph import EdgeFilter
 
+
 class BsaFormat(pyffi.object_models.xml.FileFormat):
     """This class implements the BSA format."""
     xml_file_name = 'bsa.xml'
@@ -115,6 +124,7 @@ class BsaFormat(pyffi.object_models.xml.FileFormat):
     # implementation of bsa-specific basic types
 
     class Hash(pyffi.object_models.common.UInt64):
+
         def __str__(self):
             return "0x%016X" % self._value
 
@@ -122,6 +132,7 @@ class BsaFormat(pyffi.object_models.xml.FileFormat):
             return self.__str__()
 
     class BZString(pyffi.object_models.common.SizedString):
+
         def get_size(self, **kwargs):
             return 2 + len(self._value)
 
@@ -136,6 +147,7 @@ class BsaFormat(pyffi.object_models.xml.FileFormat):
 
     class FileVersion(pyffi.object_models.common.UInt):
         """Basic type which implements the header of a BSA file."""
+
         def __init__(self, **kwargs):
             BasicBase.__init__(self, **kwargs)
 
@@ -248,7 +260,8 @@ class BsaFormat(pyffi.object_models.xml.FileFormat):
                 # morrowind
                 logger.debug("Reading file records at 0x%08X." % stream.tell())
                 self.old_files.read(stream, data=self)
-                logger.debug("Reading file name offsets at 0x%08X." % stream.tell())
+                logger.debug(
+                    "Reading file name offsets at 0x%08X." % stream.tell())
                 for old_file in self.old_files:
                     old_file._name_offset_value_.read(stream, data=self)
                 logger.debug("Reading file names at 0x%08X." % stream.tell())
@@ -266,7 +279,8 @@ class BsaFormat(pyffi.object_models.xml.FileFormat):
                 stream.seek(total_num_bytes, os.SEEK_CUR)
             else:
                 # oblivion and up
-                logger.debug("Reading folder records at 0x%08X." % stream.tell())
+                logger.debug(
+                    "Reading folder records at 0x%08X." % stream.tell())
                 self.folders.read(stream, data=self)
                 logger.debug(
                     "Reading folder names and file records at 0x%08X."
@@ -301,6 +315,6 @@ class BsaFormat(pyffi.object_models.xml.FileFormat):
             # write the file
             raise NotImplementedError
 
-if __name__=='__main__':
+if __name__ == '__main__':
     import doctest
     doctest.testmod()