+++ pyffi/qskope/__init__.py
@@ -50,6 +50,7 @@ from pyffi.formats.kfm import KfmFormat
 from pyffi.formats.dds import DdsFormat
 from pyffi.formats.tga import TgaFormat
 from pyffi.formats.egm import EgmFormat
+from pyffi.formats.esp import EspFormat
 from pyffi.formats.tri import TriFormat
 from pyffi.formats.bsa import BsaFormat
 
@@ -327,7 +328,7 @@ available via
 on <a href="http://sourceforge.net">SourceForge</a>.</p>
 <p>You are running PyFFI %s.
 The most recent version of PyFFI can always be downloaded from the
-<a href="http://sourceforge.net/project/showfiles.php?group_id=199269">
+<a href="http://sourceforge.net/projects/pyffi/files/">
 PyFFI SourceForge Project page</a>.""" % pyffi.__version__)
         # display the window
         mbox.exec_()