+++ CHANGELOG.rst
@@ -1,6 +1,32 @@
 Release 2.1.3 (in development)
 ==============================
 
+* Added toaster option to process files in archives (not yet functional).
+
+* Added toaster option to resume, by skipping existing files in the
+  destination folder.
+
+* Toaster now removes incompletely written files on CTRL-C (to avoid
+  corrupted files).
+
+* Fixed makefarnif spell (now no longer deletes vertex colors).
+
+* New spells
+
+  - fix_delunusedbones
+
+  - modify_bonepriorities
+
+  - XXX more
+
+* The niftoaster optimize spell now also includes fix_delunusedbones.
+
+* Removed unused pep8 attribute conversion code.
+
+* Toasters can now be configured from an ini file.
+
+* bhkMalleableHinge update_a_b bugfix (reported by Ghostwalker71).
+
 Release 2.1.2 (Jan 16, 2010)
 ============================